Fix the progress bar breaking after changing input once
This also exposes the fact that cancel() doesn't actually work.  I've removed the call to it from when things finish 100% (which did nothing in any case) but not the first cancel since I _will_ want to replace it with actual cancelation code if I can figure out how that'd work.